refactor codebase into seperated files and modules + fixed empty prefix
